Extended bam_list with a destruct routine, that accepts a destructor routine for the data, and destroys the data before destroying the list. This allows you to quickly and easily free resources stored in the list as well as the list.
Also, extended bam_list with foreach* routines that allow you to execute a
user-defined routine on each element in the list, and with the *_ret version,
return a new list of pointers that are the result of the user-defined routine.

This is somewhat experimental. Everything in libbam is currently considered
poorly tested and unstable, but this is especially unstable because I wrote it
during the early morning (as in last night early), which is also why I'm
committing it so late/early.

That said, the updated unit test seemed to work (though it isn't a very
thorough test).

I also added a new type, bam_foo, which is just a simple new struct used by
the bam_list_test unit test code to test out the new bam_list functionality.
It may become useful for other tests so I figured I would just add it to the
library so it was centralized, etc. Eventually, I should wrap it in a debug
macro so it is only built into the library if requested. Meh...