Whatever, dude... >_< Subversion doesn't seem to be handling the history properly...
Anyway, I'm moving include/bam/bam_* header files to just include/bam/*
because it seems the bam_ prefix is redundant with the bam subdirectory
there... This might not be permanent. I guess I'll have to see how it works
and whether there are undesirable side effects in the future. Another option
might be to get rid of the bam subdirectory, but then it would be more
difficult to control the installation, I think. Alternatively, the prefix
could be restored while holding onto the subdirectory... We'll see if there
are ill effects of the current structure the future.