Further expanding upon the string concept.
Nothing is really very well layed out yet. Added stubs for the string
functions (none of which return any values, so none of which can be used
yet) as well as adding a 'theoretical' implementation of *_as_char.
Impossible to test without the rest of the library. :P

Realized that a character type would probably be needed also to support
UTF-8 (minor oversight, but it's late and I'm tired). >_< Added some
files to support that concept, but not much thought has really gone into
it yet.

(Also fixed a minor indentation flaw in string.h)